refactor !traduzir
